pyproject.toml
[tool.black]
line-length = 88
target-version = ["py37"]
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.poetry]
name = "md5service"
version = "0.1.0"
description = ""
authors = ["Dmitry Rodionov <captfrozenberg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
aiohttp = "^3.5"
pydantic = {version = "^0.29.0",extras = ["email"]}
aioredis = "^1.2"
python-dotenv = "^0.10.3"
aiosmtplib = "^1.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
black = {version = "^18.3-alpha.0",allows-prereleases = true}
pre-commit = "^1.17"
pytest-aiohttp = "^0.3.0"
pytest-mock = "^1.10"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
